fix(collections): move functions to a global file: functions.php #339
