+++ src/Collections/composer.json
@@ -36,7 +36,7 @@
             "Aphiria\\Collections\\": "src"
         },
         "files": [
-            "src/Functions/array_list.php"
+            "src/functions.php"
         ]
     },
     "autoload-dev": {
