+++ src/Collections/src/functions.php
